refactor: rename multiband EQ module and separate tests
