+++ src/lib/recorder/mix.ts
@@ -1,5 +1,5 @@
+import type { MultibandEqParameters } from "../dsp/biquad-eq-multiband.ts";
 import { ensureBiquadEqWorklet } from "../dsp/biquad-eq-node.ts";
-import type { MultibandEqParameters } from "../dsp/multiband-eq.ts";
 import { AudioChannel } from "./audio-channel.ts";
 import type { AudioPlaybackSource } from "./audio-sources.ts";
 import { getAudioTrackSources, getTakeSources } from "./audio-sources.ts";
